Split TBAP from NSLS2-specific additions and subclass Adjust log config for APs Various cleanups
TBAP:
Add method to clear raw tag information
Make group record retrieval method set on init with import
Loading tag config is an update, not reassignment

TBPAP:
Add method to clear current cycle tags only
