+++ authz_workspace/tag_definitions.yaml
@@ -211,21 +211,8 @@ tags:
     groups:
       - name: n2sn-right-dataadmin
         role: facility_data_admin
-  test_scopes:
-    users:
-      - name: nmaytan
-        role: facility_user
 
 tag_owners:
-  test_scopes:
-    users:
-      - name: nmaytan
-  six_beamline:
-    groups:
-      - name: n2sn-right-dataadmin-six
-  sst_beamline:
-    groups:
-      - name: n2sn-right-dataadmin-sst
   n2sn_data_admin:
     groups:
       - name: n2sn-right-dataadmin
